@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.2)
 project(demo_rpc)
 
 #c++17
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 #c
 set(CMAKE_C_STANDARD 11)
 
@@ -29,17 +29,18 @@ set(HEADERS
   )
 
 set(NET_SOURCES
-    ${PROJECT_SOURCE_DIR}/shared/base/logger.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/acceptor.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/address.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/channel.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/connector.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/event_loop.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/echo_server.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/poller.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/socket.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/tcp_connection.cc
-    ${PROJECT_SOURCE_DIR}/shared/net/tcp_server.cc)
+        ${PROJECT_SOURCE_DIR}/shared/base/logger.cc
+       ${PROJECT_SOURCE_DIR}/shared/base/condition_variable.cc)
+#        ${PROJECT_SOURCE_DIR}/shared/net/acceptor.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/address.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/channel.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/connector.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/event_loop.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/echo_server.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/poller.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/socket.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/tcp_connection.cc
+#        ${PROJECT_SOURCE_DIR}/shared/net/tcp_server.cc)
 add_library(SHARED_LIB ${NET_SOURCES})
 
 add_executable(${PROJECT_NAME}
