@@ -1,7 +1,7 @@
 #ifndef DEMO_RPC_CONDITION_VARIABLE_H
 #define DEMO_RPC_CONDITION_VARIABLE_H
 #include "lock.h"
-#include <<condition_variable>
+
 class condition_variable
 {
 public:
